produce(_:)를 Beverage객체로 옮김
